+++ packages/bento-design-system/src/Table/cells.tsx
@@ -1,6 +1,5 @@
 import { ComponentProps, FunctionComponent } from "react";
 import { CellProps } from "react-table";
-import { ButtonLinkProps } from "../Button/ButtonLink";
 import {
   LocalizedString,
   Body,
@@ -9,38 +8,39 @@ import {
   IconProps,
   Label,
   Link,
-  TooltipProps,
   Children,
   Inline,
   Inset,
   Box,
+  Button,
+  Tooltip,
+  Chip,
+  IconButton,
+  IconButtonProps,
+  ButtonLink,
+  ButtonLinkProps,
 } from "..";
-import { IconButtonProps } from "../IconButton/IconButton";
 
-export function createButtonCell(Button: FunctionComponent<ButtonProps>) {
-  return function ButtonCell({
-    value: buttonProps,
-    column: { align },
-  }: CellProps<{}, Omit<ButtonProps, "size">>) {
-    return (
-      <Inline space={0} align={align} alignY="center">
-        <Button {...buttonProps} size="medium" />
-      </Inline>
-    );
-  };
+export function ButtonCell({
+  value: buttonProps,
+  column: { align },
+}: CellProps<{}, Omit<ButtonProps, "size">>) {
+  return (
+    <Inline space={0} align={align} alignY="center">
+      <Button {...buttonProps} size="medium" />
+    </Inline>
+  );
 }
 
-export function createButtonLinkCell(ButtonLink: FunctionComponent<ButtonLinkProps>) {
-  return function ButtonLinkCell({
-    value: buttonProps,
-    column: { align },
-  }: CellProps<{}, Omit<ButtonLinkProps, "size">>) {
-    return (
-      <Inline space={0} align={align}>
-        <ButtonLink {...buttonProps} size="medium" />
-      </Inline>
-    );
-  };
+export function ButtonLinkCell({
+  value: buttonProps,
+  column: { align },
+}: CellProps<{}, Omit<ButtonLinkProps, "size">>) {
+  return (
+    <Inline space={0} align={align}>
+      <ButtonLink {...buttonProps} size="medium" />
+    </Inline>
+  );
 }
 
 export function TextCell({ value, column: { align } }: CellProps<{}, LocalizedString>) {
@@ -51,58 +51,49 @@ export function TextCell({ value, column: { align } }: CellProps<{}, LocalizedSt
   );
 }
 
-export function createTextWithIconCell(Tooltip: FunctionComponent<TooltipProps>) {
-  return function TextWithIconCell({
-    value: { icon, iconPosition, text, tooltipContent },
-    column: { align },
-  }: CellProps<
-    {},
-    {
-      icon: FunctionComponent<IconProps> | null;
-      iconPosition: "left" | "right";
-      text: LocalizedString;
-      tooltipContent?: Children;
-    }
-  >) {
-    const icon_ = icon && icon({ size: 12 });
+export function TextWithIconCell({
+  value: { icon, iconPosition, text, tooltipContent },
+  column: { align },
+}: CellProps<
+  {},
+  {
+    icon: FunctionComponent<IconProps> | null;
+    iconPosition: "left" | "right";
+    text: LocalizedString;
+    tooltipContent?: Children;
+  }
+>) {
+  const icon_ = icon && icon({ size: 12 });
 
-    return (
-      <Inset space={16}>
-        <Inline space={8} alignY="center" align={align} reverse={iconPosition === "right"}>
-          {tooltipContent ? (
-            <Tooltip
-              content={tooltipContent}
-              trigger={(ref, triggerProps) => (
-                <Box ref={ref} {...triggerProps}>
-                  {icon_}
-                </Box>
-              )}
-            />
-          ) : (
-            icon_
-          )}
-          <Body size="medium">{text}</Body>
-        </Inline>
-      </Inset>
-    );
-  };
+  return (
+    <Inset space={16}>
+      <Inline space={8} alignY="center" align={align} reverse={iconPosition === "right"}>
+        {tooltipContent ? (
+          <Tooltip
+            content={tooltipContent}
+            trigger={(ref, triggerProps) => (
+              <Box ref={ref} {...triggerProps}>
+                {icon_}
+              </Box>
+            )}
+          />
+        ) : (
+          icon_
+        )}
+        <Body size="medium">{text}</Body>
+      </Inline>
+    </Inset>
+  );
 }
 
-export function createChipCell<CustomColor extends string>(
-  Chip: FunctionComponent<ChipProps<CustomColor>>
-) {
-  return function ChipCell({
-    value: chipProps,
-    column: { align },
-  }: CellProps<{}, ChipProps<CustomColor>>) {
-    return (
-      <Inset space={16}>
-        <Inline space={0} align={align} alignY="center">
-          <Chip {...chipProps} />
-        </Inline>
-      </Inset>
-    );
-  };
+export function ChipCell({ value: chipProps, column: { align } }: CellProps<{}, ChipProps>) {
+  return (
+    <Inset space={16}>
+      <Inline space={0} align={align} alignY="center">
+        <Chip {...chipProps} />
+      </Inline>
+    </Inset>
+  );
 }
 
 export function LabelCell({ value, column: { align } }: CellProps<{}, LocalizedString>) {
@@ -134,15 +125,13 @@ export function IconCell({
   );
 }
 
-export function createIconButtonCell(IconButton: FunctionComponent<IconButtonProps>) {
-  return function IconButtonCell({
-    value: iconButtonProps,
-    column: { align },
-  }: CellProps<{}, Omit<IconButtonProps, "size" | "kind" | "hierarchy">>) {
-    return (
-      <Inline space={0} align={align} alignY="center">
-        <IconButton kind="transparent" hierarchy="primary" {...iconButtonProps} size={16} />
-      </Inline>
-    );
-  };
+export function IconButtonCell({
+  value: iconButtonProps,
+  column: { align },
+}: CellProps<{}, Omit<IconButtonProps, "size" | "kind" | "hierarchy">>) {
+  return (
+    <Inline space={0} align={align} alignY="center">
+      <IconButton kind="transparent" hierarchy="primary" {...iconButtonProps} size={16} />
+    </Inline>
+  );
 }
